requirements.txt
discord.py==2.6.4
python-dotenv==1.1.1
aiohttp==3.13.0
pillow==12.0.0
aiosqlite==0.21.0

pytest==9.0.1
pytest_asyncio==1.3.0
pytest_mock==3.15.1